pyproject.toml
[project]
name = "finetune1210-thunder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "huggingface-hub>=0.36.0",
    "japanize-matplotlib>=1.1.3",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
]
